+++ vvipers/Viper.cpp
@@ -6,7 +6,7 @@ void Viper::draw(sf::RenderTarget& target, sf::RenderStates states) const {
     target.draw(m_viperGraph, states);
 }
 
-void Viper::setup(const Vec2& from, float angle, const Time& length) {
+void Viper::setup(const Vec2& from, double angle, const Time& length) {
     m_viperPhys.setup(from, angle, length);
 }
 void Viper::update(const Time& elapsedTime) {
